pdf_table_extractor/requirements.txt
streamlit
camelot-py[cv]
pandas
openpyxl
deep-translator
langdetect